pyproject.toml
[project]
name = "mcp-image"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.2.1",
    "pillow>=11.1.0",
]
